Attempt update and move to pnpm and vite
