+++ .gitignore
@@ -1,2 +1,3 @@
 release-builds/
 build/
+node_modules/
