+++ client/components/ChapterCellComponent.jsx
@@ -1,9 +1,9 @@
 import React, { PropTypes } from 'react'
-import { TableRow, TableRowColumn } from 'material-ui/Table'
-import IconButton from 'material-ui/IconButton'
-import ActionGetApp from 'material-ui/svg-icons/action/get-app'
-import ActionDelete from 'material-ui/svg-icons/action/delete'
-import CircularProgress from 'material-ui/CircularProgress'
+import { TableRow, TableRowColumn } from '@mui/material/Table'
+import IconButton from '@mui/material/IconButton'
+import ActionGetApp from '@mui/material/svg-icons/action/get-app'
+import ActionDelete from '@mui/material/svg-icons/action/delete'
+import CircularProgress from '@mui/material/CircularProgress'
 import { ipcRenderer } from 'electron'
 
 import { NOT_DOWNLOADED, DOWNLOADING, DOWNLOADED } from '../../utils/constants.js'
