+++ client/components/HeaderComponent.jsx
@@ -1,11 +1,11 @@
 import React, { PropTypes } from 'react'
-import AppBar from 'material-ui/AppBar'
-import ActionNoteAdd from 'material-ui/svg-icons/action/note-add'
-import AvLoop from 'material-ui/svg-icons/av/loop'
-import IconButton from 'material-ui/IconButton'
-import Dialog from 'material-ui/Dialog'
-import FlatButton from 'material-ui/FlatButton'
-import TextField from 'material-ui/TextField'
+import AppBar from '@mui/material/AppBar'
+import ActionNoteAdd from '@mui/material/svg-icons/action/note-add'
+import AvLoop from '@mui/material/svg-icons/av/loop'
+import IconButton from '@mui/material/IconButton'
+import Dialog from '@mui/material/Dialog'
+import FlatButton from '@mui/material/FlatButton'
+import TextField from '@mui/material/TextField'
 import { validHostname } from '../../utils/url.js'
 
 const EMPTY_TEXT = 'EMPTY_TEXT'
