+++ client/components/NoMangaComponent.jsx
