+++ package-old.json
@@ -1,0 +1,61 @@
+{
+  "name": "manga-js",
+  "version": "0.0.1",
+  "description": "Desktop manga viewer in Javascript",
+  "main": "index.js",
+  "scripts": {
+    "start": "electron .",
+    "test": "jest",
+    "build": "webpack -p",
+    "package-mac": "electron-packager . --overwrite --platform=darwin --arch=x64 --prune=true --out=release-builds",
+    "package-windows": "electron-packager . --overwrite --asar=true --platform=win32 --arch=ia32 --prune=true --out=release-builds --version-string.CompanyName=DM --version-string.FileDescription=DM --version-string.ProductName=\"Manga-JS\"",
+    "package-linux": "electron-packager . --overwrite --platform=linux --arch=x64 --prune=true --out=release-builds"
+  },
+  "author": "Darin Minamoto",
+  "license": "GNU GPL 3",
+  "devDependencies": {
+    "babel-jest": "^18.0.0",
+    "babel-loader": "^6.4.0",
+    "babel-preset-env": "^1.2.1",
+    "electron": "^1.4.13",
+    "electron-packager": "^8.5.2",
+    "jest": "^18.1.0",
+    "react-test-renderer": "^15.4.1",
+    "webpack": "^2.2.1"
+  },
+  "dependencies": {
+    "babel-core": "^6.21.0",
+    "babel-plugin-transform-object-rest-spread": "^6.20.2",
+    "babel-plugin-transform-react-jsx": "^6.8.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-register": "^6.18.0",
+    "bluebird": "^3.4.7",
+    "cheerio": "^0.22.0",
+    "cloudscraper": "^1.4.1",
+    "fs-extra": "^1.0.0",
+    "immutable": "^3.8.1",
+    "lodash": "^4.17.3",
+    "material-ui": "^0.16.5",
+    "mime-types": "^2.1.14",
+    "node-fetch": "^1.6.3",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
+    "react-measure": "^1.4.6",
+    "react-redux": "^5.0.1",
+    "react-redux-toastr": "^4.4.2",
+    "react-router": "^3.0.0",
+    "react-router-redux": "^4.0.7",
+    "react-router-scroll": "^0.4.1",
+    "react-tap-event-plugin": "^2.0.1",
+    "redux": "^3.6.0",
+    "redux-logger": "^2.7.4",
+    "redux-thunk": "^2.1.0"
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/release-builds/",
+      "/build/"
+    ]
+  }
+}
