+++ package.json
@@ -1,61 +1,33 @@
 {
   "name": "manga-js",
-  "version": "0.0.1",
-  "description": "Desktop manga viewer in Javascript",
+  "version": "1.0.0",
+  "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "electron .",
-    "test": "jest",
-    "build": "webpack -p",
-    "package-mac": "electron-packager . --overwrite --platform=darwin --arch=x64 --prune=true --out=release-builds",
-    "package-windows": "electron-packager . --overwrite --asar=true --platform=win32 --arch=ia32 --prune=true --out=release-builds --version-string.CompanyName=DM --version-string.FileDescription=DM --version-string.ProductName=\"Manga-JS\"",
-    "package-linux": "electron-packager . --overwrite --platform=linux --arch=x64 --prune=true --out=release-builds"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "Darin Minamoto",
-  "license": "GNU GPL 3",
-  "devDependencies": {
-    "babel-jest": "^18.0.0",
-    "babel-loader": "^6.4.0",
-    "babel-preset-env": "^1.2.1",
-    "electron": "^1.4.13",
-    "electron-packager": "^8.5.2",
-    "jest": "^18.1.0",
-    "react-test-renderer": "^15.4.1",
-    "webpack": "^2.2.1"
-  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "packageManager": "pnpm@10.23.0",
   "dependencies": {
-    "babel-core": "^6.21.0",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-plugin-transform-react-jsx": "^6.8.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-register": "^6.18.0",
-    "bluebird": "^3.4.7",
-    "cheerio": "^0.22.0",
-    "cloudscraper": "^1.4.1",
-    "fs-extra": "^1.0.0",
-    "immutable": "^3.8.1",
-    "lodash": "^4.17.3",
-    "material-ui": "^0.16.5",
-    "mime-types": "^2.1.14",
-    "node-fetch": "^1.6.3",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-measure": "^1.4.6",
-    "react-redux": "^5.0.1",
-    "react-redux-toastr": "^4.4.2",
-    "react-router": "^3.0.0",
-    "react-router-redux": "^4.0.7",
-    "react-router-scroll": "^0.4.1",
-    "react-tap-event-plugin": "^2.0.1",
-    "redux": "^3.6.0",
-    "redux-logger": "^2.7.4",
-    "redux-thunk": "^2.1.0"
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/material": "^7.3.5",
+    "@reduxjs/toolkit": "^2.11.0",
+    "cheerio": "^1.1.2",
+    "immutable": "^5.1.4",
+    "lodash": "^4.17.21",
+    "mime-types": "^3.0.2",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "react-measure": "^2.5.2",
+    "react-redux": "^9.2.0",
+    "react-redux-toastr": "^8.0.0",
+    "react-router-dom": "^7.9.6"
   },
-  "jest": {
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "/release-builds/",
-      "/build/"
-    ]
+  "devDependencies": {
+    "electron": "^39.2.4",
+    "vite": "^7.2.4"
   }
 }
